pyproject.toml
[project]
name = "md2reqif"
version = "0.1.0"
description = "Convert Markdown requirements to ReqIF format and back"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "garysclaw" }]
keywords = ["reqif", "requirements", "markdown", "systems-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Documentation",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "lxml>=5.0",
    "pyyaml>=6.0",
]

[project.scripts]
md2reqif = "md2reqif.cli:main"

[project.urls]
Homepage = "https://github.com/garysclaw/md2reqif"
Repository = "https://github.com/garysclaw/md2reqif"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/md2reqif"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
